requirements.txt
streamlit~=1.39.0
firebase-admin~=6.5.0
beautifulsoup4~=4.12.3
streamlit_js_eval
unidecode~=1.3.8
pyrebase4
requests~=2.32.3
pandas~=2.2.3
pillow